api outsourced in utils folder. create a component Place with toggle function
