+++ pages/_app.js
@@ -2,7 +2,6 @@ import GlobalStyle from "@/styles";
 import Head from "next/head";
 import { useEffect, useState } from "react";
 import Image from "next/image";
-import PlacesToVisitOverview from "@/components/placesToVisitOverview";
 
 function MyApp({ Component, pageProps }) {
   return <Component {...pageProps} />;
