+++ utils/api.js
@@ -1,0 +1,22 @@
+const apiKey = process.env.NEXT_PUBLIC_apiKey;
+
+export function apiGet(method, query) {
+  return new Promise(function (resolve, reject) {
+    var otmAPI =
+      "https://api.opentripmap.com/0.1/en/places/" +
+      method +
+      "?apikey=" +
+      apiKey;
+
+    if (query !== undefined) {
+      otmAPI += "&" + query;
+    }
+
+    fetch(otmAPI)
+      .then((response) => response.json())
+      .then((data) => resolve(data))
+      .catch(function (err) {
+        reject(err);
+      });
+  });
+}
